requirements.txt
Flask==2.2.2
google-cloud-datastore==2.8.3
python-dotenv
pyjwt==1.7.1